+++ terraform/modules/main/main.tf
@@ -34,6 +34,11 @@ resource "heroku_addon" "db" {
   plan = "heroku-postgresql:hobby-dev"
 }
 
+resource "heroku_addon" "logdna" {
+  app  = heroku_app.default.name
+  plan = "logdna:quaco"
+}
+
 resource "heroku_addon" "sendgrid" {
   app  = heroku_app.default.name
   plan = "sendgrid:starter"